move symmetric shader reader and writer from lib/usd/translators to lib/mayaUsd/fileio/shading
The symmetric shader reader and writer can be useful from multiple locations in
the repo, and they aren't actually translators by themselves, so this change
relocates them to the mayaUsd core with the other shading translation
infrastructure. The header files are also now installed publicly.
