+++ lib/mayaUsd/fileio/shading/symmetricShaderReader.cpp
