+++ lib/mayaUsd/fileio/shading/symmetricShaderWriter.cpp
